tier0: basetypes: Cast FALSE to int

+++ public/tier0/basetypes.h
@@ -79,7 +79,7 @@ inline T AlignValue( T val, unsigned alignment )
 
 #ifndef FALSE
 #define FALSE 0
-#define TRUE (!FALSE)
+#define TRUE ((int)(!FALSE))
 #endif
